// components/common/types/disperse.types.ts
export type TAmountInputElement = {
	amount: string | undefined;
	normalizedBigAmount: TNormalizedBN;
	status: 'pending' | 'success' | 'error' | 'none';
	isValid: boolean | 'undetermined';
	error?: string | undefined;
};

/**********************************************************************************************
 ** TDisperseInput contains the receiver and the value of the disperse operation. The UUID is
 ** used to identify the input and control it from outside. It acts as a unique reference.
 *********************************************************************************************/
export type TDisperseInput = {
	receiver: TInputAddressLike;
	value: TAmountInputElement;
	UUID: string;
};